Create seq_of_atoms.c
Hey,

I talked with the TAs today, and they said we should use other Hanson's implementations besides Atoms. Looking more carefully, it seems that sequences are perfect for our project, as the elements are allocated contiguously and dynamically.
I pushed 2 files: one represents a sequence of atoms, and one represents a sequence of sequences (so a matrix).